Removed placeholders from Troubleshooting.
I had accidently left a couple placeholders in the databases from source
section of the Troubleshooting chapter for windows and for mac. Whilst
the mac one will likely always be the same, the windows one will change
depending on where one installs it. The other question is, is anyone on
these platforms going to be compiling from source? Possibly mac, but
windows?