Split src comp sec in doc due to change to cmake.
Split the source compilation chapter into two sections, from tarball and
from git repo, due to the move to cmake. I also fixed a couple of
formatting errors in the chapter and removed the number from the title
of the install page.

The troubleshooting section on git compiliation has also been modified
to represent this change.