Updated the first few main window images.
Updated the first few images of the main window that appear in the main
text so as to keep them up to date. There are a few other images that
remain, such as the old extender images, that are not as pressing as
the changes are slight. They will need to be updated eventually though.