Rementioned the results view in the first chapter.
I had previously taken all mentions of the results view out of the first
chapter in favour of just talking about the search panel. As it is still
called the results view in the preferences panel, I put some mention of
it back in the GUI section; It now talks about the results view as being
part of the search panel. I think this both reads better than the way I
had it before and explains the interface better.