"Finnished" updating the preferences chapter.
And when I say that, there is currently no detailed information for the
kanji stroke order animation tab and no section on user data. I've left
the user data out because the page itself is incredibly verbose.

I'm not sure if I originally included much more information in the
stroke animation section but there is currently not much. I _should_ go
through and fill it out, if only for the sake of being thorough, but as
it is everything in there should be fairly obvious, especially after
reading the rest of the manual (Which is a really weak excuse, I know.)