Strip SKIP codes from kanjidic2 and remove SKIP search function
SKIP codes are not redistributable under a free software compatible
license and thus needs to be removed from kanjidic2 to allow Tagaini to
be distributed in Debian and other distributions.

Remove SKIP codes from the source and disable SKIP search functionality.
This change is made easy to revert in case we can negociate a change of
the SKIP code. If we cannot, a complete removal will ensue.