jmdict: add placeholder text for when deleted entries are shown
The JMdict is constantly evolving, and some entries are removed or
merged. While such entries don't land in the dictionary database, the
user might still have references to them in e.g. the study list after
upgrading. We could silently delete them but this behavior would be
rather hostile.

Instead the plan is to maintain a separate database of deleted entries
with basic information (e.g. only kanji and kana) across releases so the
user can look up a replacement. In some cases though, it cannot be ruled
out that even that fallback mechanism won't work, so for these cases
let's just display a placeholder message explaining the situation and
asking to manually removing the entry after reviewing it.