jmdict: stop checking for deleted/moved entries
The JMdict does not contain information about moved or merged entries
anymore, so the only thing we could do was to delete any reference to a
non-existing entry. This causes user data loss, and right now the
program should not crash anymore when dealing with non-existing entries,
so stop doing that.

Ideally we would guide to user to find a new entry that can replace the
deleted one, but this will require some extra work. Maybe we can just
point to the JMdictDB's deleted entry page for now?